requirements.txt
numpy==1.14.3
matplotlib==2.2.2
requests==2.18.4
Pillow==5.1.0
